+++ Sources/Scout/Constants/DataFormat.swift
