+++ Sources/Scout/Models/Error/ExplorerError.swift
