+++ Sources/Scout/Models/Error/PathError.swift
