+++ Sources/Scout/Models/PathParser/PathParser+Operators.swift
