+++ Sources/Scout/Models/PathParser/PathParser.swift
